server/pyproject.toml
[project]
name = "whiteops-server"
version = "1.0.0"
description = "White-Ops Enterprise API Server - AI Workforce Platform"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.12",

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",

    # Validation
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",

    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pyotp>=2.9.0",
    "cryptography>=43.0.0",

    # Cache & Queue
    "redis>=5.2.0",
    "celery[redis]>=5.4.0",

    # Storage
    "minio>=7.2.0",

    # HTTP Client
    "httpx>=0.28.0",

    # Monitoring & Logging
    "prometheus-client>=0.21.0",
    "structlog>=24.4.0",

    # Resilience
    "tenacity>=9.0.0",

    # Scheduling
    "croniter>=3.0.0",

    # Email
    "aiosmtplib>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "factory-boy>=3.3.0",
    "faker>=30.0.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "S", "B", "A", "C4"]
ignore = [
    "E501",  # line too long (existing codebase has many; enforced by formatter instead)
    "S101",  # assert in tests
    "S104",  # binding to all interfaces
    "S105",  # false positive: hardcoded password (config defaults, JWT audience/issuer)
    "S110",  # try-except-pass (intentional fallback pattern)
    "B008",  # Depends() in function defaults (standard FastAPI pattern)
    "UP042", # StrEnum migration (requires Python 3.11+ StrEnum, defer to future cleanup)
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-report=xml"

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.coverage.run]
source = ["app"]
omit = ["app/tests/*", "app/db/migrations/*"]

[tool.coverage.report]
fail_under = 50
show_missing = true
